+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * KNIFE_RACKSPACE-38 Support the -P --purge option
 * Refactored to use msg_pair method like other knife plugins with eye on eventual base class.
 * KNIFE_RACKSPACE-29 Support private network to connect to for bootstrap
+* KNIFE_RACKSPACE-40 Support for disabling host key checking
 
 ## v0.5.14
 * KNIFE_RACKSPACE-25 version bump to match knife-ec2's dependencies